scripts/requirements.txt
# requirements.txt
python-decouple == 3.8
requests == 2.32.4
beautifulsoup4 == 4.13.4
python-dotenv == 1.1.0